requirements.txt
Django
gunicorn
django-heroku
dj-database-url
django-allauth
requests
PyJWT
cryptography
python-dotenv
Pillow
daphne
channels
channels-redis
boto3==1.34.64
botocore==1.34.64
django-storages==1.14.3
